+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - 👷(docker) add arm64 platform support for image builds
 - ✨(global) add create file from template feature
 - ✨(global) add FRONTEND_CSS_URL and FRONTEND_JS_URL settings
+- ✨(backend) add a download action returning the media url
 
 ### Fixed